requirements.txt
django
gunicorn
django-heroku
django-cors-headers
google-cloud-speech==1.3.2
firebase-admin==4.0.1
google-api-python-client
oauth2client
